requirements.txt
django
django-auditlog
psycopg2-binary
